Cambio a lista desplegables en rutas de origen a destino, y traia de datos de lugares en la base de datos, metodos crud para lugares

+++ src/main/java/com/viajesglobal/repository/LugarRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LugarRepository extends JpaRepository<Lugar,Long> {
+public interface LugarRepository extends JpaRepository<Lugar,Integer> {
 }
